% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogeereader.R
\docType{package}
\name{apogeereader}
\alias{apogeereader}
\alias{apogeereader-package}
\title{Simple loading of SpectraWiz output files in R.}
\description{
Use the \code{read_spectrawiz()} function to load SpectraWiz output files into wide
format data frames in R, or the  \code{read_spectrovision()} function for
SpectroVision .csv output-files.
}
\author{
Alexander Koc
}
