% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-spectrovision-metadata.R
\name{get_spectrovision_metadata}
\alias{get_spectrovision_metadata}
\title{Extract Metadata Columns From Spectrovision Spectra}
\usage{
get_spectrovision_metadata(spectrovision_data)
}
\arguments{
\item{spectrovision_data}{a \code{data.frame} output from \code{read_spectrovision()}}
}
\value{
a \code{data.frame} \code{data.frame} of metadata associated with
}
\description{
\code{get_spectrovision_metadata()} takes a \code{data.frame} produced by \code{read_spectrovision()}
 and outputs a \code{data.frame} containing only non-wavelength columns.
}
\examples{

path_to_spectrovision_file <- spectrovision_file()
spectral_data <- read_spectrovision(path_to_spectrovision_file)
meta_data <- get_spectrovision_metadata(spectral_data)

}
\author{
Alexander Koc
}
