% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrawizreader.R
\name{read_spectrawiz}
\alias{read_spectrawiz}
\title{Read one or more SpectraWiz output files into a dataframe}
\usage{
read_spectrawiz(file, different_spectral_ranges = FALSE)
}
\arguments{
\item{file}{a vector of paths to SpectraWiz file(s)}

\item{different_spectral_ranges}{logical (TRUE/FALSE). Do the input files
contain different spectral ranges? If yes set this to TRUE. If not leave as
FALSE for faster performance. If you attempt to load files with different
spectral ranges while this is set to false, spectrawizreader will stop with
an error.}
}
\value{
a \code{data.frame} of filenames and spectra associated with the
input files in \code{file}
}
\description{
\code{read_spectrawiz()} takes a single .TRM, .IRR, .ABS, and
.SSM file or a vector of file paths as input and outputs a \code{data.frame}
containing the spectra and the file paths of the input files.
}
\examples{

# Get the path to the demo file

path_to_spectrawiz_file <- trm_file()

spectral_data <- read_spectrawiz(path_to_spectrawiz_file)

# Print the five first columns:
print(spectral_data[1,1:5])

}
\author{
Alexander Koc
}
