% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectrovision.R
\name{read_spectrovision}
\alias{read_spectrovision}
\title{Read one or more SpectraWiz output files into a dataframe}
\usage{
read_spectrovision(file)
}
\arguments{
\item{file}{a vector of paths to SpectraWiz file(s)}
}
\value{
a \code{data.frame} spectral data and any metadata associated with
input file  \code{file}
}
\description{
\code{read_spectrovision()} takes a single SpectroVision .csv file
 and outputs a \code{data.frame} containing the spectra and metadata associated
 with \code{file}
}
\examples{

# TODO: Update example

path_to_spectrovision_file <- spectrovision_file()

spectral_data <- read_spectrawiz(path_to_spectrovision_file)

# Print the first row and five first columns:
print(spectral_data[1,1:5])

}
\author{
Alexander Koc
}
