% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectrawiz.R
\docType{package}
\name{spectrawizreader}
\alias{spectrawizreader}
\alias{spectrawizreader-package}
\title{Simple loading of SpectraWiz output files in R.}
\description{
Use the \code{read_spectrawiz()} to load SpectraWiz output files into wide
format data frames in R.
}
\author{
Alexander Koc
}
