% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{spectrovision_file}
\alias{spectrovision_file}
\title{Get file path of a sample SepctroVision .csv file}
\usage{
spectrovision_file()
}
\value{
a character vector containing a file path to a sample file
}
\description{
This function returns a path to a sample file containing two
spectral reflectance readings for testing
the \code{read_spectrovision()} function.
}
\examples{

path_to_spectrovision_file <- spectrovision_file()
path_to_spectrovision_file
}
\author{
Alexander Koc
}
