% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrawizreader.R
\name{trm_file}
\alias{trm_file}
\title{Get file path of a sample SpectraWiz .TRM file}
\usage{
trm_file()
}
\value{
a character vector containing a file path to a sample file
}
\description{
This function returns a path to a sample file containing a
spectral reading made on a winter wheat canopy, which you can use to test
the \code{read_spectrawiz()} function.
}
\examples{

path_to_spectrawiz_file <- trm_file()
path_to_spectrawiz_file
}
\author{
Alexander Koc
}
